Continued working on MlpSet::Remove
Continued working on MlpSet::Remove, focusing on adding the
functions/function signature of the required functions in the data
structures used by the MlpSet, namely CuckooHashTable and
CuckooHashTableNode.

We need to support CuckooHashTable::Remove which removes a key from a
given index in the tree, and functions that allow to "delete" (clear) a
node, or to clear a single on of its children.