Handled minkey and L1/L2 caches in MlpSet::Remove
Handled the minkey modification in MlpSet::Remove, which includes going
up the tree and adjusting the minkey to the next biggest descendant of
the current node.

Also handled L1/L2 caches by removing the appropraite bit from the L1/L2
caches in case no other value exists in the DS that has the same high
bits.
This is easily computable by retrieving the successor of the removed
number and checking to see whether their high bits match.
NOTE:
There is a bug with this approach - this will only work if the removed
value is the minkey in the subrange of its high bits, if not, then this
approach might fail as there might be a smaller number (not the
successor) which shares the same high bits.
We can work around this by retrieving the successor of:
<smallest number in high bits range> - 1, and the successor of the
removed value.
If both of those fall outside of the high bits range, then we know for
certain no other value in the DS has those high bits.