Make CuckooHashTableNode::GetChildNum support 0 children
Up until now the implementation of GetChildNum always assumed the node
had more than 0 children (due to how the DS was used and the fact that
it didn't support removing elements).

Now that we might end up with 0 children, we need to support retrieving
this value.
For this purpose we use the new childrenCount variable (which can
completely replace the existing 3 bits in hash after we test it
properly).