Add readers generations array
Added the readers generation array into MlpSet.
This array contains as many entries as there are CPUs in the system, and
holds the generation each ongoing reader has started operating with.
This member is needed for us to be able to safely reclaim memory
dynamically allocated during our runtime (specifically memory allocated
for external bitmaps).