Free unused external bitmaps buffers once it is safe
Added a vector of "awaiting deallocations" to MlpSet.
This vector contains all the external bitmaps that were removed from the
data structure.
Each time MlpSet::Remove is called we go through this structure and
de-allocate every buffer that we're certain no reader an access.
We can be certain no reader can access a bitmap if all of the ongoing
readers' generation is bigger than the generation in which the pointer
to this buffer was removed from the data structure.