Removed debug assertion of key existence in QueryLCPInternal
The assertion was used to verify that after we've detected a node in the
data structure we can find it.
However this is no longer the case - a writer might've removed the item
between the time a reader observed it in the tree and the time this
assertion is made, causing needless failures in our tests.