Add benchmark_mlp
Added a module named benchmark_mlp through which we can test both
MlpRange and MapleTree using the same code.
The module merely contains a single time-diff helper function at the
moment but is already linked into both the Maple test binary as well as
the MlpIndex test binary.
This means we can easily implement tests in both C as well as C++ that
leverages this module to benchmark both modules using the same code.