Use correct bitmasks for clearing L1/L2 cache
The L1 cache contains information about the first 2 bytes of a key.
The L2 cache contains information about the first 3 bytes of a key.
The previous code used the wrong bitmask to figure out whether or not to
remove a bit, causing some bits to linger longer than needed in the
L1/L2 cache.