Add benchmarking functionality in benchmark_mlp
Added the functionality to run arbitrary single threaded workloads in
benchmark_mlp.c.
This commit also introduces workload A and B which are very simple, and
adds tests that invoke them in both the C++ project of MlpIndex as well
as the maple tree tests projects in the Linux submodule.

This commit is intended to serve as the foundation of future benchmarks
done for both projects.