benchmark: Add iterations and cpu to routine context
Added the ability to supply amount of iterations and the CPU to be used
by the routine.
The first is important for us to be able to run long tests without
generating the input on the fly (which might become the bottleneck of
the tests instead of the actual work) and the second is important for us
to make sure each of the working threads end up on a different logical
core (to prevent them from sharing the resource thus badly impacting the
benchmarks).