benchmark: workload D - measure operations per unit of time
Started implementing workload D which measures amount of operations done
in a given amount of time.
In this workload we run one writer thread and N reader threads (as
expected) and let them run for a predefined amount of time.
We count the amount of operations they've done and report them.

This workload support 2 access patterns by the writer/reader threads:
1. AllRange - all threads access the range [start, end] of the DS.
2. ExclusiveRanges - every thread access a unique range which doesn't
   overlap with other threads' ranges.

In the future we'll add a 3rd access pattern in which every thread will
access random ranges in every iteration.